outsource auth function
