+++ middleware/auth.js
@@ -1,0 +1,31 @@
+const passport = require('../config/passport')
+const helpers = require('../_helpers')
+const jwt = require('jsonwebtoken')
+const { User } = require('../models')
+
+const authenticated = (req, res, next) => {
+  passport.authenticate('jwt', { session: false }, (err, user, info) => {
+    if (err) {
+      return next(err)
+    }
+    if (!user) {
+      return res.status(401).json({ status: 'error', message: 'without jwt' })
+    }
+    req.user = user
+    return next()
+  })(req, res, next)
+}
+
+const authenticatedAdmin = (req, res, next) => {
+  if (helpers.getUser(req)) {
+    if (helpers.getUser(req).role === 'admin') {
+      req.user = user
+      return next()
+    }
+    return res.json({ status: 'error', message: 'permission denied' })
+  } else {
+    return res.json({ status: 'error', message: 'permission denied' })
+  }
+}
+
+module.exports = { authenticated, authenticatedAdmin }
