+++ routes/apis.js
@@ -14,30 +14,8 @@ const replyController = require('../controllers/api/replyController')
 const likeController = require('../controllers/api/likeController')
 const followshipController = require('../controllers/api/followshipController')
 
-/* authenticated */
-const authenticated = (req, res, next) => {
-  passport.authenticate('jwt', { session: false }, (err, user, info) => {
-    if (err) {
-      return next(err)
-    }
-    if (!user) {
-      return res.status(401).json({ status: 'error', message: 'without jwt' })
-    }
-    req.user = user
-    return next()
-  })(req, res, next)
-}
-
-const authenticatedAdmin = (req, res, next) => {
-  if (helpers.getUser(req)) {
-    if (helpers.getUser(req).role === 'admin') {
-      return next()
-    }
-    return res.json({ status: 'error', message: 'permission denied' })
-  } else {
-    return res.json({ status: 'error', message: 'permission denied' })
-  }
-}
+/* authentication */
+const { authenticated, authenticatedAdmin } = require('../middleware/auth')
 
 // * upload image *
 const uploadImage = upload.fields([
